[YS-479] refactor: 쿼리를 관리하는 useQueryParams 분리
